# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFSO API",
        "description": "API for TFSO application",
        "contact": {
            "name": "Burak",
            "url": "https://burakkaynar.com",
            "email": "burak@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tracks": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Tüm şarkıları listeler",
                "operationId": "a73400d1e2ec7667296ffa66ff67761d",
                "responses": {
                    "200": {
                        "description": "Şarkıların listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Track"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Yeni bir şarkı oluşturur",
                "operationId": "3b575ef9cd32174c62b9ce92c31a6c37",
                "requestBody": {
                    "description": "Şarkı bilgileri",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "audio_file",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Harika Şarkı"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bu şarkı çok güzel."
                                    },
                                    "audio_file": {
                                        "description": "Ses dosyası (mp3, wav vb.)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image_file": {
                                        "description": "Kapak resmi (jpg, png vb.)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_premium": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "iap_product_id": {
                                        "type": "string",
                                        "example": "premium_track_01"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Şarkı başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Track"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tracks/{id}": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Belirli bir şarkının detaylarını getirir",
                "operationId": "5b68bb9f6d22bdc9d39e8e8437cac3af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Şarkı ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Şarkı detayları",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Track"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Şarkı bulunamadı"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Bir şarkıyı günceller",
                "operationId": "4be5b91751eca9a515a6bd8430a57a0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek şarkının ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Güncellenecek şarkı bilgileri",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "audio_file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image_file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_premium": {
                                        "type": "boolean"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "_method": {
                                        "description": "Form-data ile PUT isteği göndermek için gereklidir.",
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Şarkı başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Track"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Şarkı bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Bir şarkıyı siler",
                "operationId": "613a044835d3eef38f4e536aafa377aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Silinecek şarkının ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Şarkı başarıyla silindi"
                    },
                    "404": {
                        "description": "Şarkı bulunamadı"
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Track": {
                "title": "Track",
                "description": "Track model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "audio_file": {
                        "type": "string"
                    },
                    "image_file": {
                        "type": "string"
                    },
                    "is_premium": {
                        "type": "boolean"
                    },
                    "iap_product_id": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "audio_url": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "image_url": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "audio_file_name": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        }
    },
...
}